+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/computercraft/peripheral/BoundMethod.kt
@@ -19,6 +19,11 @@ class BoundMethod(private val target: Any, val name: String, private val method:
         return method.apply(target, context, access, arguments)
     }
 
+    fun equalWithoutTarget(other: Any?): Boolean {
+        if (this === other) return true
+        if (other !is BoundMethod) return false
+        return name == other.name && method == other.method
+    }
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other !is BoundMethod) return false