+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/computercraft/peripheral/DisabledPeripheral.kt
@@ -5,4 +5,8 @@ import site.siredvin.peripheralium.computercraft.peripheral.owner.DisabledPeriph
 object DisabledPeripheral : OwnedPeripheral<DisabledPeripheralOwner>("disabled", DisabledPeripheralOwner()) {
     override val isEnabled: Boolean
         get() = true
+
+    override fun equals(other: Any?): Boolean {
+        return other === this
+    }
 }